--- src/screens/plant-chat/components/chatting-form/index.ts
@@ -1,1 +1,0 @@
-export * from './chatting-form';
