+++ src/screens/plant-chat/module/index.ts
@@ -1,2 +1,3 @@
 export * from './plant-chat-header';
+export * from './plant-chat-footer';
 export * from './plant-chat-content';
