+++ src/screens/plant-chat/module/plant-chat-footer/index.ts
@@ -1,0 +1,1 @@
+export * from './plant-chat-footer.module';
